+++ edsnlp/utils/training.py
@@ -1,14 +1,12 @@
 import os
-import sys
 import tempfile
 from enum import Enum
 from itertools import islice
 from pathlib import Path
 from random import shuffle
-from typing import IO, TYPE_CHECKING, Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, List, Optional, Union
 
 import spacy
-from rich_logger import RichTablePrinter
 from spacy.errors import Errors, Warnings
 from spacy.schemas import ConfigSchemaTraining
 from spacy.tokens import Doc, DocBin
@@ -16,7 +14,6 @@ from spacy.training.loop import train as train_loop
 from spacy.util import get_sourced_components, logger, registry, resolve_dot_names
 from thinc.api import ConfigValidationError, fix_random_seed, set_gpu_allocator
 from thinc.config import Config
-from tqdm import tqdm
 
 from edsnlp.connectors import BratConnector
 from edsnlp.utils.merge_configs import merge_configs
@@ -83,10 +80,6 @@ DEFAULT_TRAIN_CONFIG = Config().from_str(
         compound = 1.001
         t = 0.0
 
-    [training.logger]
-        @loggers = "eds.RichLogger.v1"
-        progress_bar = false
-
     [training.optimizer]
         @optimizers = "Adam.v1"
         beta1 = 0.9
@@ -110,7 +103,7 @@ DEFAULT_TRAIN_CONFIG = Config().from_str(
 )
 
 
-class DataFormat(Enum):
+class DataFormat(str, Enum):
     brat = "brat"
     standoff = "brat"
     spacy = "spacy"
@@ -119,7 +112,7 @@ class DataFormat(Enum):
 def make_spacy_corpus_config(
     train_data: Union[str, List[Doc]],
     dev_data: Union[str, List[Doc], int, float],
-    data_format: Optional[DataFormat] = None,
+    data_format: Union[Optional[DataFormat], str] = None,
     nlp: Optional[spacy.Language] = None,
     seed: int = 0,
     reader: str = "spacy.Corpus.v1",
@@ -337,109 +330,3 @@ def train(
 
     os.makedirs(output_path, exist_ok=True)
     train_loop(nlp, output_path)
-
-
-@registry.loggers("eds.RichLogger.v1")
-def console_logger(
-    progress_bar: bool = False,
-) -> Callable[
-    [spacy.Language],
-    Tuple[Callable[[Optional[Dict[str, Any]]], None], Callable[[], None]],
-]:
-    """
-    A rich based logger that renders nicely in Jupyter notebooks and console
-
-    Parameters
-    ----------
-    progress_bar: bool
-        Whether to show a training progress bar or not
-
-    Returns
-    -------
-    Tuple[Callable[[Optional[Dict[str, Any]]], None], Callable[[], None]]]
-    """
-
-    def setup_printer(
-        nlp: "Language", stdout: IO = sys.stdout, stderr: IO = sys.stderr
-    ) -> Tuple[Callable[[Optional[Dict[str, Any]]], None], Callable[[], None]]:
-
-        # ensure that only trainable components are logged
-        logged_pipes = [
-            name
-            for name, proc in nlp.pipeline
-            if hasattr(proc, "is_trainable") and proc.is_trainable
-        ]
-        eval_frequency = nlp.config["training"]["eval_frequency"]
-        score_weights = nlp.config["training"]["score_weights"]
-        score_cols = [
-            col for col, value in score_weights.items() if value is not None
-        ] + ["speed"]
-
-        fields = {"epoch": {}, "step": {}}
-        for pipe in logged_pipes:
-            fields[f"loss_{pipe}"] = {
-                "format": "{0:.2f}",
-                "name": f"Loss {pipe}".upper(),
-                "goal": "lower_is_better",
-            }
-        for score, weight in score_weights.items():
-            if score != "speed" and weight is not None:
-                fields[score] = {
-                    "format": "{0:.2f}",
-                    "name": score.upper(),
-                    "goal": "higher_is_better",
-                }
-        fields["speed"] = {"name": "WPS"}
-        fields["duration"] = {"name": "DURATION"}
-        table_printer = RichTablePrinter(fields=fields)
-
-        progress: Optional[tqdm] = None
-        last_seconds = 0
-
-        def log_step(info: Optional[Dict[str, Any]]) -> None:
-            nonlocal progress, last_seconds
-
-            if info is None:
-                # If we don't have a new checkpoint, just return.
-                if progress is not None:
-                    progress.update(1)
-                return
-
-            data = {
-                "epoch": info["epoch"],
-                "step": info["step"],
-            }
-
-            for pipe in logged_pipes:
-                data[f"loss_{pipe}"] = float(info["losses"][pipe])
-
-            for col in score_cols:
-                score = info["other_scores"].get(col, 0.0)
-                try:
-                    score = float(score)
-                except TypeError:
-                    err = Errors.E916.format(name=col, score_type=type(score))
-                    raise ValueError(err) from None
-                if col != "speed":
-                    score *= 100
-                data[col] = score
-            data["duration"] = info["seconds"] - last_seconds
-            last_seconds = info["seconds"]
-
-            if progress is not None:
-                progress.close()
-            table_printer.log(data)
-
-            if progress_bar:
-                # Set disable=None, so that it disables on non-TTY
-                progress = tqdm(
-                    total=eval_frequency, disable=None, leave=False, file=stderr
-                )
-                progress.set_description(f"Epoch {info['epoch'] + 1}")
-
-        def finalize() -> None:
-            table_printer.finalize()
-
-        return log_step, finalize
-
-    return setup_printer